requirements.txt
google-api-core==2.11.1
google-auth==2.36.0
google-cloud-bigquery==3.14.1
google-cloud-core==2.3.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
python-dotenv==1.0.1
selenium==4.26.1
setuptools==75.6.0
undetected-chromedriver==3.5.5
webdriver-manager==4.0.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.95.0
requests==2.31.0